+++ packages/ts-newskit/src/slices/content-bucket-1/index.tsx
@@ -12,10 +12,6 @@ import {
   LeadArticle,
   LeadArticleProps
 } from '../../components/slices/lead-article';
-import {
-  SliceHeader,
-  SliceHeaderProps
-} from '../../components/slices/slice-header';
 import { ArticleProps } from '../../components/slices/article';
 import { LeadStoryDivider, StackItem, BlockItem } from '../shared-styles';
 
@@ -28,14 +24,12 @@ import {
 } from '../../components/slices/shared-styles';
 
 export interface ContentBucket1Props {
-  section: SliceHeaderProps;
   leadArticle: LeadArticleProps;
   comments: CommentCardProps[];
   articles: ArticleProps[];
 }
 
 export const ContentBucket1 = ({
-  section,
   leadArticle,
   comments,
   articles
@@ -58,18 +52,6 @@ export const ContentBucket1 = ({
 
   return (
     <CustomStackLayout>
-      <FullWidthBlock>
-        <BlockItem
-          $width={{
-            xs: '100%',
-            md: '720px',
-            lg: '976px',
-            xl: '1276px'
-          }}
-        >
-          <SliceHeader {...section} />
-        </BlockItem>
-      </FullWidthBlock>
       <StackItem
         $width={{
           xs: '100%',